+++ lua/cmp_obsidian_new.lua
@@ -54,9 +54,10 @@ end
 
 source.execute = function(_, item, callback)
   local data = item.data
+  ---@type obsidian.Client
   local client = obsidian.new(data.opts)
   local dir = vim.fn.expand "%:p:h"
-  if client.opts.never_current_dir then
+  if client.opts.completion.new_notes_location == "notes_subdir" then
     dir = nil
   end
   client:new_note(data.title, data.id, dir)