+++ lua/obsidian/init.lua
@@ -137,9 +137,7 @@ obsidian.setup = function(opts)
       end
       local lines = note:frontmatter_lines(nil, frontmatter)
       vim.api.nvim_buf_set_lines(bufnr, 0, note.frontmatter_end_line and note.frontmatter_end_line or 0, false, lines)
-      if not self.opts.silent then
-        echo.info "Updated frontmatter"
-      end
+      echo.info("Updated frontmatter", self)
     end,
   })
 
@@ -261,9 +259,7 @@ client.new_note = function(self, title, id, dir)
   -- Create Note object and save.
   local note = obsidian.note.new(new_id, aliases, {}, path)
   note:save(nil, not self.opts.disable_frontmatter)
-  if not self.opts.silent then
-    echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path))
-  end
+  echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path), self)
 
   return note
 end
@@ -311,9 +307,7 @@ client.today = function(self)
   local note = obsidian.note.new(id, { alias }, { "daily-notes" }, path)
   if not note:exists() then
     note:save(nil, not self.opts.disable_frontmatter)
-    if not self.opts.silent then
-      echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path))
-    end
+    echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path), self)
   end
 
   return note
@@ -340,9 +334,7 @@ client.yesterday = function(self)
   local note = obsidian.note.new(id, { alias }, { "daily-notes" }, path)
   if not note:exists() then
     note:save(nil, not self.opts.disable_frontmatter)
-    if not self.opts.silent then
-      echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path))
-    end
+    echo.info("Created note " .. tostring(note.id) .. " at " .. tostring(note.path), self)
   end
 
   return note